+++ src/hkimjp/konpy2/routes.clj
@@ -28,15 +28,15 @@
     ;;["toggle/:e"  {:post admin/toggle-status!}]
     ]
    ["/k/" {:middleware [[wrap-defaults site-defaults] m/wrap-users]}
-    ["tasks"       {:get tasks/konpy}]
-    ["problem/:e"  {:get tasks/problem}]
-    ["answer"      {:post answers/post-answer}]
-    ["answer/:e/:p"   {:get  answers/show-answer}]
-    ["comment"     {:post comments/post-comment}]
-    ["comment/:e"  {:get  comments/show-comment}]
+    ["tasks"        {:get tasks/konpy}]
+    ["problem/:e"   {:get tasks/problem}]
+    ["answer"       {:post answers/post-answer}]
+    ["answer/:e/:p" {:get  answers/show-answer}]
+    ["comment"      {:post comments/post-comment}]
+    ["comment/:e"   {:get  comments/show-comment}]
     ;;
-    ["scores"      {:get dummy}]
-    ["stocks"      {:get dummy}]]
+    ["scores"       {:get dummy}]
+    ["stocks"       {:get dummy}]]
    ; ["/hx/" {:middleware [[wrap-defaults api-defaults] m/wrap-users]}
    ;  ["hello" {:post hx/dummy-post}]]
    ])